+++ eksCluster.ts
@@ -1,0 +1,27 @@
+import * as pulumi from "@pulumi/pulumi";
+import * as eks from "@pulumi/eks";
+import { eksVpc } from "./networking";
+
+// Grab some values from the Pulumi configuration (or use default values)
+const config = new pulumi.Config();
+const minClusterSize = config.getNumber("minClusterSize") || 3;
+const maxClusterSize = config.getNumber("maxClusterSize") || 6;
+const desiredClusterSize = config.getNumber("desiredClusterSize") || 3;
+const eksNodeInstanceType = config.get("eksNodeInstanceType") || "t3.medium";
+// Create the EKS cluster
+export const eksCluster = new eks.Cluster("pyramid-eks-cluster", {
+    // Put the cluster in the new VPC created earlier
+    vpcId: eksVpc.vpcId,
+    // Public subnets will be used for load balancers
+    publicSubnetIds: eksVpc.publicSubnetIds,
+    privateSubnetIds: eksVpc.privateSubnetIds,
+    // Change configuration values to change any of the following settings
+    instanceType: eksNodeInstanceType,
+    desiredCapacity: desiredClusterSize,
+    minSize: minClusterSize,
+    maxSize: maxClusterSize,
+    // Do not give the worker nodes public IP addresses
+    nodeAssociatePublicIpAddress: false,
+    endpointPrivateAccess: false,
+    endpointPublicAccess: true,
+});
